requirements.txt
# 对抗性图像生成项目依赖
# 使用以下命令安装: pip install -r requirements.txt

# 核心机器学习框架
tensorflow>=2.8.0

# 数值计算和数据处理
numpy>=1.21.0

# 数据可视化
matplotlib>=3.5.0

# 图像处理
Pillow>=9.0.0

# 对抗性攻击工具包
adversarial-robustness-toolbox>=1.10.0

# 网络请求（用于下载ImageNet类别索引）
requests>=2.28.0

# 可选：用于更好的中文显示支持（如果系统缺少中文字体）
# 在Windows上通常不需要，Linux/Mac可能需要安装中文字体包
# 对于Linux: sudo apt-get install fonts-wqy-microhei
# 对于Mac: brew install wqy-microhei
